Used blueprints to organize routes
