+++ users_schema.sql
@@ -4,5 +4,5 @@ CREATE TABLE users (
     user_id SERIAL PRIMARY KEY,
     username VARCHAR(50) UNIQUE NOT NULL,
     email VARCHAR(100) UNIQUE NOT NULL,
-    password VARCHAR(256) NOT NULL,
+    password VARCHAR(256) NOT NULL
 );
